Updates doc-md to support changes in format and implements table of contents.
Now reads the title for each section from the properties.yml file, rather than from the headers within the markdown file. This allows the document to keep the headers consistent throughout, rather than relying on the user to ensure consistency.

Implements a table of contents.

Adds a hook for user-defined css.

Replaces the grunt-bower-concat task with grunt-bower-task and grunt-contrib-concat to accomplish the same thing, only with more control.

Adds a bower.json file for configuring bower dependencies.