Changes API to better handle multiple directories, cleans up code.
Changes doc-md from a multi task (i.e., one grunt task needs to be run for each directory) to a regular task, which will process an array of directories. This change was necessary to allow for easier linking between the multiple guides, as well as more effective sharing of resources.

Refactors much of the html manipulation code into its own file for readability.

Refactors the doc-md task into multiple methods for readability.